+++ summingbird-scalding-test/src/test/scala/com/twitter/summingbird/scalding/ScaldingLaws.scala
@@ -53,24 +53,6 @@ import org.specs2.mutable._
   * Tests for Summingbird's Scalding planner.
   */
 
-// This is not really usable, just a mock that does the same state over and over
-class LoopState[T](init: T) extends WaitingState[T] { self =>
-  def begin = new PrepareState[T] {
-    def requested = self.init
-    def fail(err: Throwable) = {
-      println(err)
-      self
-    }
-    def willAccept(intr: T) = Right(new RunningState[T] {
-      def succeed = self
-      def fail(err: Throwable) = {
-        println(err)
-        self
-      }
-    })
-  }
-}
-
 object ScaldingLaws extends Specification {
   import MapAlgebra.sparseEquiv
